chore: upgrade RN version to 0.72.4
