+++ template/.eslintrc.js
@@ -13,7 +13,7 @@ module.exports = {
     'plugin:@typescript-eslint/eslint-recommended',
     'plugin:@typescript-eslint/recommended',
     'plugin:prettier/recommended',
-    '@react-native-community',
+    '@react-native',
   ],
   parser: '@typescript-eslint/parser',
   parserOptions: {
