+++ template/.gitignore
@@ -29,17 +29,15 @@ build/
 local.properties
 *.iml
 *.hprof
+.cxx/
+*.keystore
+!debug.keystore
 
 # node.js
 #
 node_modules/
 npm-debug.log
 yarn-error.log
-
-# BUCK
-buck-out/
-\.buckd/
-!debug.keystore
 
 # fastlane
 #
@@ -58,3 +56,9 @@ buck-out/
 # CocoaPods
 /ios/Pods/
 .nova
+/vendor/bundle/
+
+# Temporary files created by Metro to check the health of the file watcher
+.metro-health-check*
+# testing
+/coverage
