+++ template/__tests__/App-test.tsx
@@ -6,6 +6,10 @@ import 'react-native';
 import React from 'react';
 import App from '../App';
 
+
+// Note: import explicitly to use the types shiped with jest.
+import {it} from '@jest/globals';
+
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
