+++ template/package.json
@@ -1,6 +1,6 @@
 {
   "name": "boilerplate",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "private": true,
   "scripts": {
     "android": "react-native run-android",
@@ -16,8 +16,8 @@
     "@react-navigation/native": "^6.1.4",
     "@react-navigation/native-stack": "^6.9.10",
     "@reduxjs/toolkit": "^1.9.3",
-    "react": "18.0.0",
-    "react-native": "0.69.7",
+    "react": "18.2.0",
+    "react-native": "0.72.4",
     "react-native-device-info": "^10.4.0",
     "react-native-safe-area-context": "^4.5.0",
     "react-native-screens": "^3.20.0",
@@ -26,31 +26,38 @@
   },
   "devDependencies": {
     "reactotron-redux": "^3.1.3",
-    "@babel/core": "^7.12.9",
-    "@babel/runtime": "^7.12.5",
-    "@react-native-community/eslint-config": "^2.0.0",
-    "@tsconfig/react-native": "^2.0.2",
+    "@babel/core": "^7.20.0",
+    "@babel/preset-env": "^7.20.0",
+    "@babel/runtime": "^7.20.0",
+    "@react-native/eslint-config": "^0.72.2",
+    "@react-native/metro-config": "^0.72.11",
+    "@tsconfig/react-native": "^3.0.0",
     "@types/jest": "^26.0.23",
+    "@types/react": "^18.0.24",
     "@types/react-native": "^0.69.3",
     "@types/react-test-renderer": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^5.52.0",
     "@typescript-eslint/parser": "^5.52.0",
-    "babel-jest": "^26.6.3",
+    "babel-jest": "^29.2.1",
     "babel-plugin-module-resolver": "4.1.0",
-    "eslint": "^8.34.0",
+    "eslint": "^8.19.0",
     "eslint-config-google": "^0.14.0",
     "eslint-import-resolver-babel-module": "^5.3.2",
     "eslint-import-resolver-typescript": "^3.5.3",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-react": "^7.32.2",
-    "jest": "^26.6.3",
+    "jest": "^29.2.1",
     "metro-react-native-babel-preset": "^0.70.3",
     "react-test-renderer": "18.0.0",
     "reactotron-react-native": "^5.0.3",
-    "typescript": "^4.4.4"
+    "typescript": "^4.8.4",
+    "prettier": "^2.4.1"
   },
   "jest": {
     "preset": "react-native",
     "moduleFileExtensions": ["ts", "tsx", "js", "jsx", "json", "node"]
+  },
+  "engines": {
+    "node": ">=16"
   }
 }
