+++ template/tsconfig.json
@@ -32,5 +32,6 @@
       "@utils/*": ["./src/utils/*"]
     }
   },
-  "exclude": ["node_modules", "babel.config.js", "metro.config.js", "jest.config.js"]
+  "exclude": ["node_modules", "babel.config.js", "metro.config.js", "jest.config.js"],
+  "extends": "@tsconfig/react-native/tsconfig.json"
 }
